fix(rating): crate new component for rating and import his
