+++ frontend/src/productPages/productpage.module.css
@@ -41,38 +41,12 @@
     justify-content: center;
     margin-top: 2%;
     margin-bottom: 10%;
-    gap: 24px;
-    font-size: 2.3rem;
-    color: var(--gold);
 }
 
-.starsContainer {
-    display: flex;
-    gap: 6px;
-}
-
-.halfStar {
-    color: transparent;
-    background: linear-gradient(to right, gold 50%, #ddd 50%);
-    background-clip: text;
-    -webkit-background-clip: text;
+.productRatingContainer{
     font-size: 2rem;
-}
-
-.filledStar {
-    color: var(--gold);
-    font-size: 2rem;
-}
-
-.emptyStar {
-    color: var(--header-grey);
-    font-size: 2rem;
-}
-
-.ratingText {
     font-weight: bold;
-    font-size: 2rem;
-    color: var(--text-color);
+
 }
 
 .productButtonContainer {
@@ -124,5 +98,13 @@
     height: auto;
     font-size: 20px;
     font-weight: bold;
+}
+
+.productNotFound {
+    margin-top: 20%;
+    text-align: center;
+    justify-content: center;
+    font-weight: bold;
+    font-size: 3rem;
 }
